+++ event-sourcing/event-source-api/src/main/java/uk/gov/justice/services/eventsourcing/discovery/EventSubscriptionDiscoveryBean.java
@@ -6,6 +6,7 @@ import static javax.transaction.Transactional.TxType.REQUIRES_NEW;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.discovery.EventDiscoveryRepository;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.discovery.StreamPosition;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -28,15 +29,21 @@ public class EventSubscriptionDiscoveryBean {
     private EventDiscoveryConfig eventDiscoveryConfig;
 
     @Transactional(REQUIRES_NEW)
-    public List<StreamPosition> discoverNewEvents(final Optional<UUID> latestKnownEventId) {
+    public DiscoveryResult discoverNewEvents(final Optional<UUID> latestKnownEventId) {
 
         final int batchSize = eventDiscoveryConfig.getBatchSize();
-        if(latestKnownEventId.isPresent()) {
-            return eventDiscoveryRepository.getLatestStreamPositions(
-                    latestKnownEventId.get(),
-                    batchSize);
-        }
 
-        return eventDiscoveryRepository.getLatestStreamPositions(ZEROTH_EVENT_NUMBER, batchSize);
+        final long firstEventNumber = latestKnownEventId
+                .map(eventDiscoveryRepository::getEventNumberFor)
+                .orElse(ZEROTH_EVENT_NUMBER);
+
+        return eventDiscoveryRepository.getLatestEventIdAndNumberAtOffset(firstEventNumber, batchSize)
+                .map(latestEvent -> {
+                    final List<StreamPosition> streamPositions = eventDiscoveryRepository.getLatestStreamPositionsBetween(
+                            firstEventNumber,
+                            latestEvent.eventNumber());
+                    return new DiscoveryResult(streamPositions, Optional.of(latestEvent.id()));
+                })
+                .orElseGet(() -> new DiscoveryResult(Collections.emptyList(), Optional.empty()));
     }
 }